Transient resource system in C#

+++ Managed/Ferrum3D/Ferrum.Samples.Textures/Program.cs
@@ -224,7 +224,7 @@ namespace Ferrum.Samples.Textures
                 builder.BindVertexBuffer(0, vertexBuffer);
                 builder.BindIndexBuffer(indexBuffer);
                 builder.BeginRenderPass(renderPass, framebuffers[i], Colors.MediumAquamarine);
-                builder.DrawIndexed(indexData.Length, 1, 0, 0, 0);
+                builder.DrawIndexed((uint)indexData.Length, 1, 0, 0, 0);
                 builder.EndRenderPass();
             }
